+++ package.json
@@ -34,7 +34,8 @@
   "license": "MIT",
   "dependencies": {
     "graphql": "^16.8.1",
-    "graphql-yoga": "^5.1.1"
+    "graphql-yoga": "^5.1.1",
+    "@graphql-tools/schema": "^10.0.0"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20240112.0",
