+++ src/types.d.ts
@@ -1,0 +1,4 @@
+interface Env {
+  DEEPSEEK_API_KEY: string;
+  ALLOWED_ORIGINS?: string;
+}
